requirements.txt
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
boto3==1.28.60
botocore==1.31.60
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
cryptography==41.0.4
cssselect2==0.7.0
DateTime==5.2
Django==4.2.5
django-brutebuster2==0.2.0
holiday==1.0.0
holidays==0.34
html5lib==1.1
idna==3.4
jmespath==1.0.1
lxml==4.9.3
oscrypto==1.3.0
Pillow==10.0.1
psycopg==3.1.12
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.21
pyHanko==0.20.1
pyhanko-certvalidator==0.24.1
pyotp==2.9.0
pypdf==3.16.2
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
qrcode==7.4.2
razorpay==1.4.1
reportlab==3.6.13
requests==2.31.0
s3transfer==0.7.0
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
typing_extensions==4.8.0
tzlocal==5.0.1
uritools==4.0.2
urllib3==1.26.17
webencodings==0.5.1
xhtml2pdf==0.2.11
zope.interface==6.0
